+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ## [0.0.1] - 2023-01-?
 
 ### Added
+- Created HashRouter class to handle client-side routing
+- Created ResourceHandler class to handle the resources(eventListeners, promises, intervals, timeouts) of the components
+- Created BaseComponent class as an abstract class for creating reusable components
 - index.html as the template HTML file
 - index.js as the entry point for the application
 - Created initial project structure